requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyascii==0.3.2
    # via wagtail
asgiref==3.6.0
    # via django
beautifulsoup4==4.11.2
    # via wagtail
boto3==1.28.20
    # via -r requirements.in
botocore==1.31.20
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2023.5.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
cryptography==40.0.2
    # via
    #   django-anymail
    #   pyjwt
defusedxml==0.7.1
    # via
    #   python3-openid
    #   willow
dj-database-url==2.0.0
    # via -r requirements.in
django==4.2.4
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-anymail
    #   django-braces
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-phonenumber-field
    #   django-recaptcha
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-allauth==0.55.0
    # via -r requirements.in
django-anymail[mailgun]==10.1
    # via -r requirements.in
django-braces==1.15.0
    # via -r requirements.in
django-filter==22.1
    # via wagtail
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-phonenumber-field[phonenumberslite]==7.1.0
    # via -r requirements.in
django-recaptcha==3.0.0
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
django-taggit==4.0.0
    # via
    #   -r requirements.in
    #   wagtail
django-treebeard==4.7
    # via wagtail
djangorestframework==3.14.0
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.2.0
    # via willow
fontawesomefree==6.4.0
    # via -r requirements.in
gevent==23.9.0.post1
    # via -r requirements.in
greenlet==2.0.2
    # via gevent
gunicorn==21.2.0
    # via -r requirements.in
html5lib==1.1
    # via wagtail
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
l18n==2021.3
    # via wagtail
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.2
    # via wagtail
packaging==23.1
    # via gunicorn
phonenumberslite==8.13.21
    # via django-phonenumber-field
pillow==10.0.0
    # via
    #   -r requirements.in
    #   pillow-heif
    #   wagtail
pillow-heif==0.12.0
    # via willow
psycopg==3.1.10
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
python-dateutil==2.8.2
    # via botocore
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
requests==2.30.0
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.1
    # via boto3
sentry-sdk==1.29.2
    # via -r requirements.in
six==1.16.0
    # via
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
telepath==0.3
    # via wagtail
typing-extensions==4.5.0
    # via
    #   dj-database-url
    #   psycopg
urllib3==1.26.15
    # via
    #   botocore
    #   django-anymail
    #   requests
    #   sentry-sdk
wagtail==5.1.1
    # via
    #   -r requirements.in
    #   wagtailcodeblock
wagtailcodeblock @ git+https://github.com/fourfridays/wagtailcodeblock.git@d6b4f3ba51b9ac3f1eaac9a8fe9ac12565eb8e37
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
whitenoise==6.5.0
    # via -r requirements.in
willow[heif]==1.6.1
    # via
    #   -r requirements.in
    #   wagtail
zope-event==5.0
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
